EVEREST-392 fix PMM secret management
In order to keep auto-generating the rest of the cluster secret's data
in PXC we can't create a secret with just the PMM API key. We need to
wait for the secret to be autogenerated and then update it with the PMM
API key.

In PSMDB we can create the secret right away with just the PMM API key
and the rest of the secret's data will still be auto-generated.

In PG we shall not use the separate secret as the users. This can be
created before the DB because the operator doesn't change this secret.