EVEREST-397 fix PSMDB restore deadlock
By changing the DB cluster state to "restoring" when a restoration is in
progress we introduced a deadlock where the restore isn't able to be
updated because it's waiting for the DB to be ready and the DB won't
ever be ready while the restore isn't ready. In this commit we also
accept the DB restoring state to reconcile the restore CR which removes
the deadlock.