EVEREST-633 use dedicated namespace for monitoring configs
In order to scrape k8s metrics we need to deploy the Victoria Metrics
operator. The system namespace has an operator group with multiple
target namespaces which isn't supported by the VMOperator, hence it
shall be deployed onto a separate monitoring namespace with an operator
group with single target namespace. Since we need the VM operator and
corresponding configurations in a separate monitoring namespace, it's a
logical decision to also move the MonitoringConfigs onto this namespace
instead of duplicating data around.