EVEREST-768 fix dbc restart race condition
When pausing a PG cluster, the PG operator sets the STS replicas to 0
and sets the state to paused. Since the pods don't immediately go down,
if we unpause the cluster the stateful set replicas will go up again and
some pods might not be restarted.
Solution: wait for all pods to go down before reinitializing the DBC.