Fix MySQL PITR restore to new cluster (#317)
When performing a PITR for MySQL, the created pxc-restore object always assumes its own bucket as the data source. This will not work when we want to PITR restore to a new cluster, as it should use the source cluster's bucket and not its own.
This commit updates the PXC restore code so that it uses the same bucket as the one used for the backup. This way it works for same-cluster and cross-cluster PITR restore.

Signed-off-by: Mayank Shah <mayank.shah@percona.com>