Reverse the behaviour of storage finalizers
Earlier, we always removed the S3 finalizer from Backups,
and conditionally added it from the Everest API, if specified by the user.

This commit reverses this behaviour, i.e, we ensure the S3 finalizer is always present.
With this change, S3 is always cleaned up, unless the user chooses to skip it from the
Everest API.

Signed-off-by: Mayank Shah <mayank.shah@percona.com>