EVEREST-1354 set PXC resource requests to equal the limits
In order to provide a predictable behavior we shall set the resource's
requests to the same value as the limits.

If you specify a limit for a resource, but do not specify any request,
and no admission-time mechanism has applied a default request for that
resource, then Kubernetes copies the limit you specified and uses it as
the requested value for the resource.
This means that we don't need to explicitly set the requests in Everest
in order for them to be the same as the limits. However, since we used
to set the requests before v1.2.0 we need to continue explicitly setting
them otherwise, if users tried to scale existing DBs by increasing the
resource limits, the requests wouldn't be reconciled and would continue
assuming the old values.